broken orbit into creation and handling
